pyproject.toml
[project]
name = "ekimo-assignment"
version = "0.1.0"
description = "A FastAPI-based web application for the Ekimo developer assignment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.12.1",
    "dotenv>=0.9.9",
    "faker>=37.4.0",
    "fastapi>=0.116.0",
    "httpx>=0.28.1",
    "pandas>=2.3.1",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary>=2.9.7",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.23",
    "sqlalchemy-orm>=1.2.10",
    "typer>=0.16.0",
    "uvicorn[standard]>=0.24.0",
]

[dependency-groups]
dev = [
    "approvaltests>=14.7.1",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-approvaltests>=0.2.4",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.2",
    "sqlalchemy[mypy]>=2.0.41",
]

[tool.mypy]
ignore_missing_imports = true
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.ruff]
line-length = 88

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "venv",
]

[tool.ruff.lint]
select = [
    "A",    # flake8-builtins
    "ARG",  # flake8-unused-arguments
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "PT",   # flake8-pytest-style
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
]


[tool.ruff.lint.mccabe]
max-complexity = 10
